Разделено париснг csv и сохранение фактов
