--- Depmon.Server.Collector/IDataSave.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-using System.IO;
-
-namespace Depmon.Server.Collector
-{
-    public interface IDataSave
-    {
-        void Save(IList<Stream> dataList);
-    }
-}
